Add my agenda items
Many of these items were largely or almost entirely the work of others, and I'm just trying to drive consensus on them. I'm championing a lot of things; if anyone on TC39 wants to co-champion some of them or take some over from me, please contact me (littledan@igalia.com) and we can work together on that.